transport: Add disable_pkcs11 build pragma
This allows disabling pkcs11 when importing the transport package,
improving the portability of the resulting binary.
